backend-fastapi/pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich~=0.9.2",
    "aiomysql~=0.3.2",
    "fastapi~=0.128.0",
    "loguru~=0.7.3",
    "passlib~=1.7.4",
    "pydantic~=2.12.5",
    "pydantic-settings~=2.12.0",
    "tortoise-orm~=0.25.3",
    "uvicorn~=0.40.0",
]

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."